Move components into shared components/admin folder
