+++ front/app/components/admin/ModeratorUserSearch/messages.ts
@@ -1,0 +1,12 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  addModerators: {
+    id: 'app.components.UserSearch.addModerators',
+    defaultMessage: 'Add',
+  },
+  searchUsers: {
+    id: 'app.components.UserSearch.searchUsers',
+    defaultMessage: 'Type to search users...',
+  },
+});
