+++ front/app/containers/Admin/projects/project/permissions/Project/ProjectManagement/index.tsx
@@ -8,14 +8,14 @@ import {
 } from '@citizenlab/cl2-component-library';
 import styled from 'styled-components';
 
+import ModeratorList from 'components/admin/ModeratorList/ModeratorList';
+import UserSearch from 'components/admin/ModeratorUserSearch';
 import SeatInfo from 'components/admin/SeatBasedBilling/SeatInfo';
 import { Section } from 'components/admin/Section';
 
 import { FormattedMessage, useIntl } from 'utils/cl-intl';
 
 import messages from './messages';
-import ModeratorList from './ModeratorList';
-import UserSearch from './UserSearch';
 
 const ModeratorSubSection = styled(Section)`
   margin-bottom: 30px;
